requirements.txt
Django==2.2.17
djangorestframework==3.12.2
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
